Part 3
1. Create CurrencyManipulator class that stores all the information about the selected currency.
The class must have:
1.1 String currencyCode — currency code, i.e. USD. It consists of three letters.
1.2 Map<Integer, Integer> denominations is a map of (denomination, quantity) pairs.
To see which currency is associated with the manipulator, add a getter for currencyCode.
Obviously, the manipulator can't function without the currency's name, so add a constructor with this parameter and initialize currencyCode.

2. There may be several currencies, so we need a factory to create and store the manipulators.
Create a CurrencyManipulatorFactory class with the static getManipulatorByCurrencyCode(String currencyCode) method.
In this method, we'll create the relevant manipulator if it doesn't already exist, or return a previously created one.
Ignore case when searching for a currency manipulator.
Where do you think the best place to store all the manipulators would be? Here's a little hint: the field should be called map.

Make it impossible to create CurrencyManipulatorFactory objects.