Part 4
1. Let's decide what operation to start with.
Think about it. The ATM doesn't have money yet, so we can't test INFO and WITHDRAW.
We'll start with the DEPOSIT operation.
We read the currency code from the console. Then we read the denomination and number of banknotes. And then add them to the manipulator.

2. To read the currency code, add a static String requestCurrencyCode() method to ConsoleHelper.
This method should prompt the user to enter a currency code. Verify that the code contains 3 characters.
If the currency code is invalid, then inform the user and repeat.
If the currency code is valid, then convert the code to uppercase and return it.

3. To read the denomination and number of banknotes, add a static String[] method getTwoValidNumbers(String currencyCode) to ConsoleHelper.
This method should prompt the user to enter two positive integers.
The first number is the denomination. The second is the number of banknotes.
There is no validation of the denomination. That is, 1200 is an acceptable denomination.
If the currency code is invalid, then inform the user and repeat.

Example input:

200 5
4. In the CurrencyManipulator class, create a void addAmount(int denomination, int count) method that adds banknotes in the specified denomination and quantity.

5. It's time to see the application in action.
In the main method, write the code to implement the logic described in item 1. By the way, to avoid problems with testing the server, add "Locale.setDefault(Locale.ENGLISH);" as the first line in the main method.
Let's start it, debug it, and check it out