Part 6
To refactor the code according to the Command pattern, you need to extract several logical blocks of code.
For now, we have two such blocks: 1) code for the DEPOSIT operation, 2) code for the INFO operation.
This code is in the main method. We need to get rid of it.
Now you need to make it so the user is the one who chooses the operation to perform.

1. Add a static Operation getAllowableOperationByOrdinal(Integer i) method to the Operation enum
It should return one of the enum's fields: 1 - INFO, 2 - DEPOSIT, 3 - WITHDRAW, 4 - EXIT;
Throw an IllegalArgumentException if invalid data is entered.

2. In the ConsoleHelper class, implement the logic for a static Operation requestOperation() method.
Ask the user to choose an operation.
If the user enters 1, execute the INFO command, 2 - DEPOSIT, 3 - WITHDRAW, 4 - EXIT;
Use the method described in item 1.
Handle any exceptions and repeat the request for an operation.