Part 8
It's time to whip our main method into shape. It already has a bunch of stuff that shouldn't be there.

1. Move the logic from main to DepositCommand and InfoCommand.
So what's happened to main? We have a loop, in which we ask the user for an operation and then call a method on the CommandExecutor.
And so on forever... You need to come up the exit condition for the loop.
Correct the loop to make it a do-while. The exit condition is the EXIT operation.

2. Let's run the program and top up the account with "EUR 100 2" and "USD 20 6". Then we'll look at INFO.
It's not clear what to do, since we've created 2 manipulators: the first for EUR, the second for USD.
Let's improve the logic of InfoCommand. We need to display the balance for each manipulator.

2.1. In the CurrencyManipulatorFactory class, create a static getAllCurrencyManipulators() method that returns a Collection of all manipulators.
You stored all the manipulators in a map, right? If not, then refactor.
2.2. In InfoCommand, display "<currency code> - <total amount of money for the selected currency>".
Let's run the program and top up the account with "EUR 100 2" and "USD 20 6". Then we'll look at INFO.

Does everything work correctly?

EUR - 200
USD - 120
Excellent!
3. Run the program and immediately request the INFO operation as the first operation. Nothing is displayed? That's no good.
Add a boolean hasMoney() method to the manipulator. It should indicate whether any banknotes have been added.

4. In InfoCommand, use the method in item 3 and display "No money available." if there is no money in the ATM.